consumer-topic/requirements.txt
pika
minio
boto3
openai-whisper==20231117
transformers
torch
torchaudio
influxdb-client
spacy

# Additional requirements for BERTopic
bertopic
umap-learn
hdbscan
scikit-learn
pandas
tqdm
